+++ backend/lm_backend/api/routes/configurations.py
@@ -2,6 +2,7 @@ from typing import List, Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 
+from lm_backend.api.cruds.configuration import ConfigurationCRUD
 from lm_backend.api.cruds.generic import GenericCRUD
 from lm_backend.api.models.configuration import Configuration
 from lm_backend.api.models.feature import Feature
@@ -9,6 +10,7 @@ from lm_backend.api.models.license_server import LicenseServer
 from lm_backend.api.models.product import Product
 from lm_backend.api.schemas.configuration import (
     ConfigurationCompleteCreateSchema,
+    ConfigurationCompleteUpdateSchema,
     ConfigurationCreateSchema,
     ConfigurationSchema,
     ConfigurationUpdateSchema,
@@ -22,7 +24,7 @@ from lm_backend.permissions import Permissions
 router = APIRouter()
 
 
-crud_configuration = GenericCRUD(Configuration, ConfigurationCreateSchema, ConfigurationUpdateSchema)
+crud_configuration = ConfigurationCRUD(Configuration, ConfigurationCreateSchema, ConfigurationUpdateSchema)
 crud_product = GenericCRUD(Product, ProductCreateSchema, ProductUpdateSchema)
 crud_feature = GenericCRUD(Feature, FeatureCreateSchema, FeatureUpdateSchema)
 crud_license_server = GenericCRUD(LicenseServer, LicenseServerCreateSchema, LicenseServerUpdateSchema)
@@ -149,15 +151,87 @@ async def read_configuration(
 )
 async def update_configuration(
     configuration_id: int,
-    configuration_update: ConfigurationUpdateSchema,
+    configuration_update: ConfigurationCompleteUpdateSchema,
     secure_session: SecureSession = Depends(secure_session(Permissions.CONFIG_EDIT)),
 ):
-    """Update a configuration in the database."""
-    return await crud_configuration.update(
-        db_session=secure_session.session,
-        id=configuration_id,
-        obj=configuration_update,
-    )
+    """
+    Update a configuration in the database.
+
+    If there are features in the payload, they'll be updated if they have an id,
+    or they will be created if the id is None.
+
+    If there are license servers in the payload, they'll updated if they have an id,
+    or they will be created if the id is None.
+
+    All resources related to the configuration that aren't present in the payload
+    will be deleted.
+    """
+    if configuration_update.features is not None:
+        await crud_configuration.delete_features(
+            db_session=secure_session.session,
+            configuration_id=configuration_id,
+            payload=configuration_update,
+        )
+
+        for feature_update in configuration_update.features:
+            if feature_update.id:
+                await crud_feature.update(
+                    db_session=secure_session.session,
+                    id=feature_update.id,
+                    obj=FeatureUpdateSchema(**feature_update.dict(exclude={"id"})),
+                )
+            else:
+                feature_obj = {
+                    "name": feature_update.name,
+                    "product_id": feature_update.product_id,
+                    "config_id": configuration_id,
+                    "reserved": feature_update.reserved,
+                }
+                await crud_feature.create(
+                    db_session=secure_session.session, obj=FeatureCreateSchema(**feature_obj)
+                )
+
+    if configuration_update.license_servers is not None:
+        await crud_configuration.delete_license_servers(
+            db_session=secure_session.session,
+            configuration_id=configuration_id,
+            payload=configuration_update,
+        )
+
+        for license_server_update in configuration_update.license_servers:
+            if license_server_update.id:
+                await crud_license_server.update(
+                    db_session=secure_session.session,
+                    id=license_server_update.id,
+                    obj=LicenseServerUpdateSchema(**license_server_update.dict(exclude={"id"})),
+                )
+            else:
+                license_server_obj = {
+                    "config_id": configuration_id,
+                    "host": license_server_update.host,
+                    "port": license_server_update.port,
+                }
+                await crud_license_server.create(
+                    db_session=secure_session.session, obj=LicenseServerCreateSchema(**license_server_obj)
+                )
+
+    if any(value for _, value in configuration_update.dict(exclude={"features", "license_servers"}).items()):
+        return await crud_configuration.update(
+            db_session=secure_session.session,
+            id=configuration_id,
+            obj=ConfigurationUpdateSchema(
+                **configuration_update.dict(exclude={"features", "license_servers"})
+            ),
+        )
+    elif configuration_update.features or configuration_update.license_servers:
+        return await crud_configuration.read(
+            db_session=secure_session.session, id=configuration_id, force_refresh=True
+        )
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Please provide a valid field to update configuration.",
+        )
 
 
 @router.delete(